% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{corr.matrix}
\alias{corr.matrix}
\title{Compute pair-wise correlations and hypothesis test}
\usage{
corr.matrix(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
list with two tables: corrs (correlations), pvals (p-values)
}
\description{
Computes pair-wise correlations between variables in a dataframe df
Tests H0: rho = 0 vs H1: rho!=0 by computing p-values.
}
\examples{
library(sda1)
corr.matrix(bike[,c("temp","hum","windspeed")])
}
