% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{pred_interval_reg}
\alias{pred_interval_reg}
\title{Plot prediction intervals for simple linear regression \cr}
\usage{
pred_interval_reg(
  lmobject,
  conf_int_line = T,
  pred_interval = T,
  level = 0.95,
  ngrid = 100
)
}
\arguments{
\item{lmobject}{a fitted regression model from \code{lm}.}

\item{conf_int_line}{if TRUE, then conf intervals for regression line are plotted.}

\item{pred_interval}{if TRUE, then prediction intervals are plotted.}

\item{level}{confidence level, default is level = 0.95}

\item{ngrid}{number of grid points over which the intervals are computed}
}
\value{
data frame with confidence and prediction intervals over xGrid
}
\description{
Plot prediction intervals for simple linear regression \cr
}
\examples{
library(regkurs)
lmfit = lm(mpg ~ hp, data = mtcars)
res = pred_interval_reg(lmfit)
}
