% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regsimulate}
\alias{regsimulate}
\title{Simulate from a linear regression model}
\usage{
regsimulate(
  n,
  betavect,
  sigma_eps,
  intercept = TRUE,
  covdist = "normal",
  rho_x = 0,
  sigma_x = rep(1, length(betavect) - intercept)
)
}
\arguments{
\item{n}{the number of observations in the simulated dataset.}

\item{betavect}{a vector with regression coefficients
c(beta_0,beta_1,...beta_k). First element is intercept if \code{intercept = TRUE}}

\item{sigma_eps}{standard deviation of the error terms, epsilon.}

\item{intercept}{if \code{TRUE} an intercept is added to the model.}

\item{covdist}{distribution of the covariates. Options: \code{'normal'} or \code{'uniform'}.}

\item{rho_x}{correlation among the covariates. Same for all covariate pairs.}

\item{sigma_x}{vector with standard deviation of the covariates.}
}
\value{
dataframe with simulated data (y, X1, X2, ..., XK) (no intercept included).
}
\description{
Simulates a dataset with \code{n} observation from the linear regression model
\deqn{y = \beta_0 + \beta_1 x_1 + \ldots + \beta_k x_k + \epsilon, \epsilon \sim N(0, \sigma_\epsilon^2)}{y = \beta_0 + \beta_1 * x_1 + ... + \beta_k * x_k + \epsilon, \epsilon ~ N(0, \sigma_ \epsilon^2)}
with covariates (x) simulated from a normal distribution with the same correlation \code{rho_x} \cr
between all pairs of covariates. Covariate \eqn{x_j}{x_j} has standard deviation \code{sigma_x[j]}. \cr
Alternatively the covariate can follow a uniform distribution.
}
\examples{
library(sda1)
simdata <- regsimulate(n = 500, betavect = c(1, -2, 1, 0), sigma_eps = 2)
lmfit <- lm(y ~ X1 + X2 + X3, data = simdata)
regsummary(lmfit, anova = F)
}
