% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{residuals4in1}
\alias{residuals4in1}
\title{Residual analysis mimicing the 4-in-1 plots from Minitab}
\usage{
residuals4in1(lm_object)
}
\arguments{
\item{lm_object}{a fitted regression model from \code{lm}.}
}
\description{
Plots:
\enumerate{
\item Normal QQ-plot
\item Residuals vs fitted values
\item Histogram and normal density fit
\item Residuals vs order.
}
}
\examples{
library(sda1)
fit = lm(mpg ~ hp, data = mtcars)
residuals4in1(fit)
}
