% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda1-package.R
\docType{package}
\name{sda1-package}
\alias{sda1}
\alias{sda1-package}
\title{sda1: R-paket för kursen Statistik och dataanalys I vid SU}
\description{
Funktioner för grundläggande statistik, inkl regression.
}
\author{
\strong{Maintainer}: Mattias Villani \email{mattias.villani@gmail.com} (\href{https://orcid.org/0000-0003-2786-2519}{ORCID})

}
\keyword{internal}
